recipes-core: Add udev-toradex-rules recipe
These rules make sure that a symlink is created for all i2c and tty
devices. This symlinks  are named like on the Toradex carrier boards.
This makes it easy to identify which device in software is matching
the hardware pins.

Additional non-standard devices that are not named clearly on the
carrier board are left out to prevent confusion.

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>


(cherry picked from commit 47ef64803548f73a0eda753c5d3d3f7f315af606)