u-boot-toradex: imx: fix imx-boot recipe use case
If the final boot binary is assembled by the imx-boot recipe then
the uuu_bootloader_tag class must not be included.
This is the case if BOOT_PROVIDES_BOOT_CONTAINER is set to 0 and
the machine is based on a member of the i.MX 8 or i.MX 9 series.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 9342b8ac6d78caa4c5de8b4181976d13c73c3bfd)