btnxpuart: increase sleep between modules load
When the btnxpuart firmware loaded by mwifiex is not ready yet,
btnxpuart attempts to load its firmware during the probe, resulting
in "FW download timeout" failure.

Such behavior is often observed on the Verdin-AM62 Solo variant
with the following messages:
[   12.008351] Bluetooth: Core ver 2.22
[   12.043831] NET: Registered PF_BLUETOOTH protocol family
[   12.043865] Bluetooth: HCI device and connection manager initialized
[   12.043898] Bluetooth: HCI socket layer initialized
[   12.043910] Bluetooth: L2CAP socket layer initialized
[   12.043969] Bluetooth: SCO socket layer initialized
[   20.357223] Bluetooth: BNEP (Ethernet Emulation) ver 1.3
[   20.357312] Bluetooth: BNEP socket layer initialized
[   74.726669] Bluetooth: hci0: FW Download Timeout.

To address this issue, increase the sleep between modules-load
to allow mwifiex enough time to load the firmware before the
btnxpuart probe.

Related-to: ELB-5924
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>