u-boot: distro-boot-script: fix set console baudrate condition
With the introduction of [1] in U-Boot, the lack of match on the
`setexpr sub` command is not considered an error. As a result, the
baudrate is not set in console var leading AM62 platforms to not boot.

Fix this by comparing the `_res` string with `console` or if it is empty
instead and add baudrate if they equal.
This ensures the functionality before and after [1].

[1] commit 4c7363068651 ("cmd: setexpr: fix no matching string in gsub return empty value")

Related-to: ELB-6000
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>