alsa: ucm2: apalis/colibri-imx*: fix playback volumes and capture configuration
Fix the bootsequence by stablishing values that were previously used by
the BSP 6, making sure the audio out of the box experience increase: the
audio playback volume needs to be in a reasonable default value and the
microphone should be working and unmuted.

These changes also fixes failed LAVA tests for audio-simple and
audio-extended.

Related-to: ELB-5712
Fixes: a1808656c011 ("alsa-ucm-conf-tdx: Add ALSA UCM support for Apalis and Colibri SoMs")
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>