recipes-core: udev: add ath11k_pci for aquila-am69
The Aquila AM69 platform utilizes the ath11k_pci Linux driver, which enumerates
the wireless interface with inconsistent naming (e.g., wlP3p1s0) compared to
other Toradex modules. This discrepancy occurs because the current udev rules
do not apply to the ath11k_pci driver on Aquila AM69.

This change adds ath11k_pci to the driver list in the udev rules, ensuring the
interface is consistently enumerated as wlan0 on Aquila AM69, aligning it with
the naming convention used on other Toradex modules.

Related-to: ELB-6137
Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>