linux-toradex-upstream: aarch32: depend on lz4-native explicitely
The aarch32 config sets LZ4 as the kernel compresion.
Add the lz4 tool to depends to not rely on the allowed
hosttools.

This fixes that Linux stable changed to use lz4 as the LZ4
compression tool which is not available in all build setups.

Fixes: 1b08915fda2a ("linux-toradex-mainline: change to use LZ4 compression instead of LZO")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 63487ee6619863e0d426d2452aedb0ca18cdc1b6)