alsa-ucm-conf-tdx: syncronize Toradex files with mainline alsa-ucm-conf
Some Toradex boards were merged into alsa-ucm-conf upstream project, and
with that some modifications were asked to be done during the review
process.

As of today alsa-ucm-conf recipe from oe-core is on version 1.2.11,
which does not include our boards. Therefore, propagate the changes to
our local recipe and apply the same logic to the boards we haven't yet
upstreamed (aquila and smarc).

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>