fs-init: remove recipe
This recipe is related to older imgaes, where Tezi wasn't available.
This script was needed because it wasn't possible to know the block
device size, so a minimal size was assumed and, on first boot, the
rootfs partition was resized to the full device size and the rootfs was
extended to fill the full size.

This is not needed anymore, today this is handled by Tezi, therefore
remove it.

Related-to: ELB-6306
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>