libusbgx: move recipe to meta-toradex-demos
The libusbgx bbappend is not BSP-specific and should not reside in the
BSP layer. Keeping it here means the bbappend will be applied to all
builds using this layer, even though it was previously protected by the
tdx-distro override.

Let's move the bbappend to the meta-toradex-demos layer, without
requiring the override.

Related-to: ELB-6287
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>