u-boot-toradex: set default value of CONFIG_SPL_LOAD_FIT_ADDRESS
Add a patch defining the default value for CONFIG_SPL_LOAD_FIT_ADDRESS.
This default value introduced upstream (with commit 4f5c48d5fdab)
matches the values already being employed on the verdin-imx8m[mp] and,
with the present commit, the same one is going to be used with the
toradex-smarc-imx8mp. This is required for building signed bootloaders.

Related-to: TOR-3887
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>