udev-toradex-rules: move USB UDC gadget rule to meta-toradex-demos
The udev rule calling gadget-start when a UDC appears only makes sense
if gadget-start is actually installed. This logic is also more relevant
to our reference images than our core BSP.

Therefore, install this rule as part of the libusbgx recipe instead of
as a default here. This libusbgx bbappend can be found in
meta-toradex-demos.

Related-to: ELB-6355
Fixes: b1838fd ("udev-toradex-rules: call gadget-start from udev")
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>