image_type_tezi: fix raw rootfs image size info
This fixes an issue with a "raw" root filesystem image, typically for a squashfs or erofs,
in which the "get_uncompressed_size" function is called on a file that
doesn't exist, causing the "uncompressed_size" field to be reported as "0".
Because of this, the progress bar in the Tezi interface is at 100% during
the whole flashing process, making progress bar information irrelevant.

This fix really reports the size of the raw image file being used,
making the flashing progress bar relevant again.

Signed-off-by: Michael Opdenacker <michael.opdenacker@rootcommit.com>