linux-firmware: add sx-pceax and sx-pcebe board files for aquila family
Currently, the required board files are not included in the upstream
linux-firmware repo. As a consequence, the integration for the
Aquila-AM69 was done in the meta-toradex-ti layer, but with the
introduction of the Aquila-iMX95, which has the same WiFi solution,
it should be moved to the common layer.

Related-to: ELB-6361
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>
(cherry picked from commit 8ee99bca5c9faa98b6a8a6a2917241cb57209102)