u-boot-toradex: add patch fixing colibri imx7 fdtfile in preboot
Add a patch that fixes the PREBOOT configuration in the defconfigs of
the Colibri iMX7 NAND and eMMC modules. This is necessary because the
same TEZI U-Boot recovery binary is used for both the NAND and eMMC
SoMs.

This is necessary to have the Colibri iMX7 eMMC SoMs function with TEZI
since we consolidated the bootcmd scripts in 'meta-toradex-tezi'.

Related-to: TEI-1301
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>
(cherry picked from commit 1019eff9776c80bf5f60fcff066ab8020f6e7aae)