linux-firmware: cleanup recipe
Prepare for future changes by cleaning up this recipe. The iMX95 and
AM69 specific do_install appends share code that can be separated for
clarity and maintainability, do so.

PACKAGE_ARCH cannot be set as the base recipe linux-firmware inherits
the allarch class. When building for multiple machines where some of
them are the Aquila AM69 or Aquila iMX95, this will lead to rebuilding
the linux-firmware package when switching machine, because the
PACKAGE_ARCH is set to all.

Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>